Completed Register and Register TB. Output tested in GTKWave. (#5)
* Completed Register and Register TB. Output tested in GTKWave.

* Delete Register_ah.vcd

Will not include .vcd files in pull requests.

* Added comma inside comment of Register.v line 1.

Thanks for noting the missing comma.  I noticed this was missing in the starter code and added the comma for testing and removed it afterwards, but I think including inside the comment tag and moving the comment tag up makes it easier for testing.  That was part of why I thought I might include the .vcd file to save the trouble of others missing the comma, etc.  This is better.

* Comma moved, num_iter updated

Moved comma inside comment of Register_tb.v line 61.  Updated ++ num_iter; to num_iter = num_iter + 1; of Register_tb.v line 129.